uparser/requirements.txt
Django==4.1.7
django-q==1.3.9
google_auth==2.16.2
google_api_python_client==2.82.0
gunicorn==20.1.0
pandas==1.5.3
psycopg2-binary==2.9.5
python-dotenv==1.0.0
python_telegram_bot==20.1
